extract legacy legend keys transformation to funnel utils
