+++ frontend/src/scenes/insights/insightLogic.ts
@@ -64,6 +64,7 @@ import { dayjs, now } from 'lib/dayjs'
 import { isInsightVizNode } from '~/queries/utils'
 import { userLogic } from 'scenes/userLogic'
 import { globalInsightLogic } from './globalInsightLogic'
+import { transformLegacyHiddenLegendKeys } from 'scenes/funnels/funnelUtils'
 
 const IS_TEST_MODE = process.env.NODE_ENV === 'test'
 const SHOW_TIMEOUT_MESSAGE_AFTER = 15000
@@ -671,27 +672,11 @@ export const insightLogic = kea<insightLogicType>([
         hiddenLegendKeys: [
             (s) => [s.filters],
             (filters) => {
-                const hiddenLegendKeys: TrendsFilterType['hidden_legend_keys'] = {}
                 if (isFilterWithHiddenLegendKeys(filters) && filters.hidden_legend_keys) {
-                    for (const [key, value] of Object.entries(filters.hidden_legend_keys)) {
-                        // Transform pre-#12113 funnel series keys to the current more reliable format.
-                        // Old: `${step.type}/${step.action_id}/${step.order}/${breakdownValues.join('_')}`
-                        // New: breakdownValues.join('::')
-                        // If you squint you'll notice this doesn't actually handle the .join() part, but that's fine,
-                        // because that's only relevant for funnels with multiple breakdowns, and that hasn't been
-                        // released to users at the point of the format change.
-                        const oldFormatMatch = key.match(/\w+\/.+\/\d+\/(.+)/)
-                        if (oldFormatMatch) {
-                            // Don't override values for series if already set from a previously-seen old-format key
-                            if (!(oldFormatMatch[1] in hiddenLegendKeys)) {
-                                hiddenLegendKeys[oldFormatMatch[1]] = value
-                            }
-                        } else {
-                            hiddenLegendKeys[key] = value
-                        }
-                    }
+                    return transformLegacyHiddenLegendKeys(filters.hidden_legend_keys)
                 }
-                return hiddenLegendKeys
+
+                return {}
             },
         ],
         filtersKnown: [
